feat: [SRM-14389]: refactoring
