+++ src/modules/85-cv/pages/health-source/HealthSourceDrawer/component/customiseHealthSource/__tests__/CustomiseHealthSource.utils.test.ts
@@ -5,8 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-import { HealthSourceTypes } from '@cv/pages/health-source/types'
-import { createHealthsourceList, getSelectedProductInfo } from '../CustomiseHealthSource.utils'
+import { createHealthsourceList } from '../CustomiseHealthSource.utils'
 import { RowData, healthSourcesPayload } from './CustomiseHealthSource.mock'
 
 describe('Test Util functions', () => {
@@ -18,15 +17,5 @@ describe('Test Util functions', () => {
     RowData.healthSourceList[1].name = 'Splunk dev 14'
     // Updating existing healthsource
     expect(createHealthsourceList(RowData, healthSourcesPayload)).toEqual([...RowData.healthSourceList])
-  })
-
-  test('Test CreateHealthsourceList when selected product is ElasticSearch', () => {
-    const selectedProduct = HealthSourceTypes.Elk
-    expect(getSelectedProductInfo(selectedProduct)).toEqual(HealthSourceTypes.ElasticSearch_Logs)
-  })
-
-  test('Test CreateHealthsourceList when selected product is HealthSourceTypes SumologicMetrics', () => {
-    const selectedProduct = HealthSourceTypes.SumologicMetrics
-    expect(getSelectedProductInfo(selectedProduct)).toEqual(HealthSourceTypes.SumologicMetrics)
   })
 })
